refactor(core): migrate exec and computer use runtime owners
